% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ate}
\alias{ate}
\title{Estimate average treatment effects using a causal forest}
\usage{
ate(fit, ...)
}
\arguments{
\item{fit}{A trained \code{\link[grf]{causal_forest}} object}

\item{...}{Additional parameters to be passed to
\code{\link[grf]{average_treatment_effect}}}
}
\value{
An instance of the \code{ate} class, which is essentialy a one-row
tibble with columns \code{estimate} and \code{std.err}
}
\description{
A thin wrapper around \code{\link[grf]{average_treatment_effect}}. Defines
the \code{ate} class so that confidence intervals and plots can be created
using generics.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 cf <- causal_forest(X, Y, W)
 a <- ate(cf, target.sample = 'treated')
}
}
