% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_example.R
\name{build_causal_model}
\alias{build_causal_model}
\title{Build a simple toy-model capable of predictions under interventions, on a
simple dataset.}
\usage{
build_causal_model(data)
}
\arguments{
\item{data}{A data.frame as generated by build_data(n), with at least
treatment variable A, confounder L, prognostic variable P, and outcome Y}
}
\value{
A GLM for the outcome Y with independent variables treatment A and
  prognostic variable P. It has been fitted using IP weighting, conditional
  on confounder L.
}
\description{
Build a simple toy-model capable of predictions under interventions, on a
simple dataset.
}
\examples{
df_dev <- build_data()
causal_model <- build_causal_model(df_dev)
summary(causal_model)
# Compare this to the 'naive' model
naive_model <- glm(Y ~ A + P, family = "binomial", data = df_dev)
summary(naive_model)
# In the naive model, treatment increases the risk due to the confounding of L.
# This has been adjusted for appropriately in the causal model.
}
