% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_example.R
\name{build_data}
\alias{build_data}
\title{Simulate some simple data, in which treatment A is a binary point treatment
and outcome Y is binary. Variable L acts as a confounder on treatment and
outcome. Variable P additionally predicts the outcome. Counterfactual
outcomes under both treatment options are generated (Y0 and Y1). The observed
outcome Y is consistent with the counterfactual outcome under the observed
treament. Intented to be used as a toy example together with build_causal_model()}
\usage{
build_data(n)
}
\arguments{
\item{n}{Number of rows}
}
\value{
A data.frame, where confounder L and prognostic variable P has
  normal distribution. Treatment A is a binary variable, with probability of
  the inverse logit of L. The counterfactual outcome Y0 is a binary variable
  with probability of the inverse logit of 0.5 + L + 1.25 * P - 0.6*A.
  Outcome Y is the observed outcome under the realized treatment A.
}
\description{
Simulate some simple data, in which treatment A is a binary point treatment
and outcome Y is binary. Variable L acts as a confounder on treatment and
outcome. Variable P additionally predicts the outcome. Counterfactual
outcomes under both treatment options are generated (Y0 and Y1). The observed
outcome Y is consistent with the counterfactual outcome under the observed
treament. Intented to be used as a toy example together with build_causal_model()
}
\examples{
df_dev <- build_data(1000)
}
