% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_eval.R
\name{cf_eval}
\alias{cf_eval}
\title{Create a causal forest evaluation object}
\usage{
cf_eval(fit, Xdat, preds = NULL)
}
\arguments{
\item{fit}{A trained \code{\link[grf]{causal_forest}} object}

\item{Xdat}{The data frame \code{fit} was trained on. It may include
categorical covariates.}

\item{preds}{Output from \code{\link[grf]{predict.causal_forest}}. If
omitted, it will be called directly to generate OOB predictions and
corresponding variance estimates. This may be time consuming.}
}
\value{
A \code{cf_eval} object, which is essentially just a list with the
following elements:
\describe{
\item{\code{Xdat}}{The same as the input \code{Xdat} but converted to a \link[tibble:tibble-package]{tibble}}
\item{\code{res}}{A \link{results} object}
\item{\code{varimp}}{A \link{varimp} (variable importance) object}
}
}
\description{
Create a causal forest evaluation object
}
\examples{
\dontrun{
 require(grf)

 Xdat <- subset(cfex, select = -c(W, Y))
 X <- make_contrasts(Xdat, 'fct')
 cf <- causal_forest(X, cfex$Y, cfex$W)

 cf_eval(cf, Xdat)
}
}
