% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfeval-package.R
\docType{package}
\name{cfeval-package}
\alias{cfeval}
\alias{cfeval-package}
\title{cfeval: Causal Forest Evaluation and Visualization}
\description{
Provides some functions to facilitate diagnostics
    and evaluation of causal forest models as implemented in the "grf" R
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ensley-nexant/cfeval}
  \item Report bugs at \url{https://github.com/ensley-nexant/cfeval/issues}
}

}
\author{
\strong{Maintainer}: John Ensley \email{jensley@nexant.com}

}
\keyword{internal}
