% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfex}
\alias{cfex}
\title{An example dataset to use in causal forests}
\format{
A data frame with 2000 rows and 13 columns
}
\usage{
cfex
}
\description{
An example dataset to use in causal forests
}
\keyword{datasets}
