% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFscore.R
\name{CFscore}
\alias{CFscore}
\title{Assess counterfactual performance of a model capable of predictions under
interventions}
\usage{
CFscore(data, model, predictions, Y, propensity_formula, ip, A, treatments)
}
\arguments{
\item{data}{A data.frame on which the model is to be validated.}

\item{model}{A glm (lm?) which can make predictions under interventions which
has treatment as a predictor, or a list of glm's for each specified
treatment in the treatments variable. Either model or predictions must be
given.}

\item{predictions}{A numeric vector of predictions under intervention, or a
list of numeric vectors of predictions under different interventions as
specified by treatments variable. Either model or predictions must be
given.}

\item{Y}{A character string indicating the name of the observed outcome
column in data, or a numeric vector of the observed outcomes.}

\item{propensity_formula}{A formula used to estimate the inverse-probability
weights for the validation data. Treatment variable should be on the left
hand side, all confounders on the right hand side. It is possible that
there is a different set of confounders in the validation dataset compared
to the model-development dataset. Either this or ipw must be given}

\item{ip}{The inverse probabilty weights for the validation data. Can be
either string indicating the name of the ip column in the validation data,
or a numeric vector of ip-weights.}

\item{A}{A character string indicating the name of the realized treatment
column in data. Must be given if ipw's are specified. It is automatically
inferred from propensity_formula if given.}

\item{treatments}{A treatment level or a list of all treatment levels for
which the counterfactual perormance measures should be evaluated.}
}
\value{
A list of performance measures (Brier, Observed/Expected, AUC) of
  the model on observed data (naive) and counterfactual data, where each
  subject is assigned to both treatment options.
}
\description{
Assess counterfactual performance of a model capable of predictions under
interventions
}
\examples{
causal_model <- glm(
  Y ~ A + P,
  family = "binomial",
  data = df_dev,
  weights = ip_weights(df_dev, A ~ L)
)
CFscore(df_val, causal_model, "Y", A ~ L, treatments = list(0,1))
}
