% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{confint.ate}
\alias{confint.ate}
\title{Confidence intervals for causal forest ATEs}
\usage{
\method{confint}{ate}(object, parm, level = 0.9, ...)
}
\arguments{
\item{object}{An \code{\link{ate}} object}

\item{parm}{Not used}

\item{level}{The desired confidence level}

\item{...}{Additional arguments (ignored)}
}
\value{
A one-row tibble with two columns giving lower and upper confidence
bounds for the estimated ATE. The column names will be \eqn{(1-level)/2}\%
and \eqn{1 - (1-level)/2}\% (by default, \verb{5.0\%} and \verb{95.0\%}).
}
\description{
Computes confidence intervals for the average treatment effect estimated by a
\code{\link[grf]{causal_forest}} model.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 cf <- causal_forest(X, Y, W)
 a <- ate(cf)
 confint(a)
}
}
