% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_fit.R
\name{estimate_subgroup_ate}
\alias{estimate_subgroup_ate}
\title{Find difference in ATE between subgroups}
\usage{
estimate_subgroup_ate(fit, subgroup, level = 0.9)
}
\arguments{
\item{fit}{A trained causal forest object from
\code{\link[grf]{causal_forest}}}

\item{subgroup}{A logical vector indicating whether an observation belongs to
the subgrouping}

\item{level}{The width of the reported confidence interval}
}
\value{
A one-row tibble with the following columns: \describe{
\item{\code{in_est}, \code{in_se}}{The ATE estimate and corresponding
standard error among observations for which \code{subgroup = TRUE}}
\item{\code{out_est}, \code{out_se}}{The ATE estimate and corresponding
standard error among observations for which \code{subgroup = FALSE}}
\item{\code{diff_est}}{The difference in ATE: \code{in_est} -
\code{out_est}} \item{\code{diff_ci_low}, \code{diff_ci_high}}{The
confidence interval around \code{diff_est}}}
}
\description{
One heuristic for evaluating whether a causal forest has detected
heterogeneity is to group observations into high- and low-CATE subgroups,
estimate the ATE for each group, and determine whether the difference is
significantly different than zero.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 cf <- causal_forest(X, Y, W)
 results <- tidy_cf(cf)

 # partition into high and low CATE subgroups
 estimate_subgroup_ate(cf, results$cate > median(results$cate))
}
}
