% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_weights.R
\name{ipt_weights}
\alias{ipt_weights}
\title{Get the inverse probability of treatment weights}
\usage{
ipt_weights(data, propensity_formula)
}
\arguments{
\item{data}{a data.frame}

\item{propensity_formula}{a formula with on the l.h.s. the treatment variable and
on the right hand side the confounders.}
}
\value{
a numeric vector of the IP weights
}
\description{
Get the inverse probability of treatment weights
}
