% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_contrasts}
\alias{make_contrasts}
\title{One-hot encode categorical covariates}
\usage{
make_contrasts(X, vars)
}
\arguments{
\item{X}{A data frame of covariates. May contain both numeric and categorical
features together.}

\item{vars}{A character vector of covariates to one-hot encode.}
}
\value{
A numeric matrix
}
\description{
The \code{\link[grf]{causal_forest}} algorithm requires inputs to be numeric
matrices. This function uses \code{\link[stats]{model.matrix}} to create such
a matrix from a data frame of covariates.
}
\details{
No levels are dropped during the one-hot encodeing; a covariate with \eqn{n}
levels will turn into \eqn{n} binary columns.

No intercept term is returned in the matrix.
}
\examples{
\dontrun{
 Xdat <- subset(cfex, select = -c(W, Y))
 make_contrasts(Xdat, 'fct')
}
}
