% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFscore.R
\name{observed_score}
\alias{observed_score}
\title{Assess performance of predictions for realized treatment values.}
\usage{
observed_score(data, model, predictions, Y)
}
\arguments{
\item{data}{A data.frame on which the model is to be validated}

\item{model}{A glm}

\item{predictions}{A numeric vector of predictions. Either this or data AND
model must be given.}

\item{Y}{A numeric vector of observed outcomes, or a character string
indicating the name of the observed outcomes in the data.}
}
\value{
A list of performance measures (Brier, Observed/Expected, AUC) of
  the model on observed data
}
\description{
Assess performance of predictions for realized treatment values.
}
\examples{
model <- glm(
  Y ~ A + P,
  family = "binomial",
  data = df_dev,
)
observed_score(data = df_val, model = model, Y = "Y")
}
