% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{plot.ate}
\alias{plot.ate}
\title{Plot a causal forest ATE}
\usage{
\method{plot}{ate}(x, level = NULL, ...)
}
\arguments{
\item{x}{An \code{\link{ate}} object}

\item{level}{The desired confidence level}

\item{...}{Additional arguments (ignored)}
}
\value{
A ggplot2 plot object
}
\description{
Plots the average treatment effect estimated by a
\code{\link[grf]{causal_forest}} model.
}
\details{
If \code{level = NULL}, the default, no confidence bands will be plotted.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 cf <- causal_forest(X, Y, W)
 a <- ate(cf)
 plot(a)
}
}
\seealso{
Other plotting methods: 
\code{\link{plot.cf_eval}()},
\code{\link{plot.results}()},
\code{\link{plot.tuning_output}()}
}
\concept{plotting methods}
