% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_eval.R
\name{plot.cf_eval}
\alias{plot.cf_eval}
\title{Visualize a causal forest evaluation object}
\usage{
\method{plot}{cf_eval}(x, kind, ...)
}
\arguments{
\item{x}{A \code{\link{cf_eval}} object}

\item{kind}{The type of plot to create}

\item{...}{Additional arguments passed to subsequent plot functions}
}
\value{
A plot
}
\description{
Many different figures and diagnostic plots can be created. Specify the
desired plot using the \code{kind} argument.
}
\details{
Possible options for \code{kind} are

\describe{
\item{\code{cate}}{A density plot of estimated conditional average
treatment effects, i.e. the causal forest predictions. The most
straightforward way to look for treatment effect heterogeneity.}

\item{\code{bias}}{A histogram of each observation's contribution to the
overall bias of the model, relative to a simple difference in means.}

\item{\code{propensities}}{A histogram of fitted propensities. The causal
forest requires the assumption that we cannot deterministically tell the
treatment status of an individual given its covariates. In other words, none
of the propensity scores should be near zero or one.}

\item{\code{balnum, balcat}}{After accounting for propensity, covariate
distributions should be balanced between treated and control observations.
\code{balnum} plots overlaid histograms, one for treated and one for
control, of each numeric covariate. \code{balcat} plots stacked bar charts
of the proportions of each categorical covariate.}

\item{\code{catecovar}}{A scatter plot of estimated CATEs as a function of a
certain covariate.}
}
}
\examples{
\dontrun{
 require(grf)

 Xdat <- subset(cfex, select = -c(W, Y))
 X <- make_contrasts(Xdat, 'fct')
 cf <- causal_forest(X, cfex$Y, cfex$W)

 cfe <- cf_eval(cf, Xdat)
 plot(cfe, kind = 'cate')
 plot(cfe, kind = 'balnum', covar = c('V1', 'V2', 'V3'))
}
}
\seealso{
Other plotting methods: 
\code{\link{plot.ate}()},
\code{\link{plot.results}()},
\code{\link{plot.tuning_output}()}
}
\concept{plotting methods}
