% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.results}
\alias{plot.results}
\title{Visualize a causal forest results object}
\usage{
\method{plot}{results}(x, kind, ...)
}
\arguments{
\item{x}{A \code{\link{results}} object}

\item{kind}{The type of plot to create}

\item{...}{Additional arguments passed to subsequent plot functions}
}
\value{
A plot
}
\description{
Certain plots only require a \code{\link{results}} object and not the other
components of a \code{\link{cf_eval}} object. These can be created from
\code{results} only.
}
\details{
Possible options for \code{kind} are:
\describe{
\item{\code{cate}}{A density plot of estimated conditional average
treatment effects, i.e. the causal forest predictions. The most
straightforward way to look for treatment effect heterogeneity.}

\item{\code{bias}}{A histogram of each observation's contribution to the
overall bias of the model, relative to a simple difference in means.}

\item{\code{propensities}}{A histogram of fitted propensities. The causal
forest requires the assumption that we cannot deterministically tell the
treatment status of an individual given its covariates. In other words, none
of the propensity scores should be near zero or one.}
}
}
\examples{
\dontrun{
 require(grf)

 Xdat <- subset(cfex, select = -c(W, Y))
 X <- make_contrasts(Xdat, 'fct')
 cf <- causal_forest(X, cfex$Y, cfex$W)

 cfe <- cf_eval(cf, Xdat)
 plot(cfe$res, kind = 'bias')
}
}
\seealso{
Other plotting methods: 
\code{\link{plot.ate}()},
\code{\link{plot.cf_eval}()},
\code{\link{plot.tuning_output}()}
}
\concept{plotting methods}
