% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_output.R
\name{plot.tuning_output}
\alias{plot.tuning_output}
\title{Plot causal forest tuning results}
\usage{
\method{plot}{tuning_output}(x, param = NULL, ...)
}
\arguments{
\item{x}{The \code{tuning_output} object that is part of a
trained \code{\link[grf]{causal_forest}} object}

\item{param}{The parameter to plot on the horizontal axis. Defaults to the
first parameter in \code{tuning_output$params}}

\item{...}{Additional arguments (ignored)}
}
\value{
A ggplot2 plot object
}
\description{
If hyperparameter tuning was performed during the causal forest fit, this
function will plot the error as a function of one of the tuned parameters.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
 cf <- causal_forest(X, Y, W, tune.parameters = 'min.node.size')

 plot(cf$tuning_output)
}
}
\seealso{
Other plotting methods: 
\code{\link{plot.ate}()},
\code{\link{plot.cf_eval}()},
\code{\link{plot.results}()}
}
\concept{plotting methods}
