% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_fit.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot a histogram of estimated bias}
\usage{
plot_bias(results, ...)
}
\arguments{
\item{results}{The output from \code{\link{tidy_cf}}}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{geom_histogram}}.}
}
\value{
A ggplot2 plot object
}
\description{
Plot a histogram of estimated bias
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 # fit the model and extract results
 cf <- causal_forest(X, Y, W)
 results <- tidy_cf(cf)

 plot_bias(results, bins = 30)
}
}
