% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_fit.R
\name{plot_calibration_test}
\alias{plot_calibration_test}
\title{Plot calibration test results}
\usage{
plot_calibration_test(clbr, ci = 0.9)
}
\arguments{
\item{clbr}{The output of \code{\link[grf]{test_calibration}}}

\item{ci}{The size of the confidence intervals. Default is 90\%}
}
\value{
A ggplot2 plot object
}
\description{
Display the results of \code{\link[grf]{test_calibration}} in graphical
format.
}
\details{
From the \code{grf} documentation:

The forest summary function \code{\link[grf]{test_calibration}} can be used
to asses a forest's goodness of fit. A coefficient of 1 for
\code{mean.forest.prediction} suggests that the mean forest prediction is
correct and a coefficient of 1 for \code{differential.forest.prediction}
suggests that the forest has captured heterogeneity in the underlying signal.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
 cf <- causal_forest(X, Y, W)

 clbr <- test_calibration(cf)
 plot_calibration_test(clbr, ci = 0.95)
}
}
