% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_output.R
\name{plot_cate}
\alias{plot_cate}
\title{Plot a histogram of estimated CATEs}
\usage{
plot_cate(results)
}
\arguments{
\item{results}{The output from \code{\link{tidy_cf}}}
}
\value{
A ggplot2 plot object
}
\description{
Plot a histogram of estimated CATEs
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
 cf <- causal_forest(X, Y, W)
 results <- tidy_cf(cf)

 plot_cate(results)
}
}
