% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_output.R
\name{plot_predictions_vs_covariate}
\alias{plot_predictions_vs_covariate}
\title{Plot estimated CATEs versus a covariate}
\usage{
plot_predictions_vs_covariate(dat, res, covariate, level = NULL)
}
\arguments{
\item{dat}{The original training data}

\item{res}{The output from \code{\link{tidy_cf}}}

\item{covariate}{The covariate to plot on the horizontal axis}

\item{level}{The desired confidence level. If \code{NULL}, no confidence
levels are plotted.}
}
\value{
A ggplot2 plot object
}
\description{
Plot estimated CATEs versus a covariate
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
 cf <- causal_forest(X, Y, W)
 results <- tidy_cf(cf)
 dat <- dplyr::bind_cols(as.data.frame(X), results)
 dat$a <- sample(letters[1:3], size = n, replace = T)

 plot_predictions_vs_covariate(dat, 'V1', level = 0.9)
 plot_predictions_vs_covariate(dat, 'a', level = 0.9)
}
}
