% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_overlap.R
\name{plot_propensities}
\alias{plot_propensities}
\title{Plot a histogram of fitted propensities}
\usage{
plot_propensities(results, ...)
}
\arguments{
\item{results}{A trained causal forest object from \code{\link[grf]{causal_forest}}}

\item{...}{Additional arguments to be passed to \code{\link[ggplot2]{geom_histogram}}}
}
\value{
A ggplot2 plot object
}
\description{
The causal forest requires the assumption that we cannot deterministically tell the
treatment status of an individual given its covariates. In other words, none of the
propensity scores should be near zero or one.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
 cf <- causal_forest(X, Y, W)

 plot_propensities(cf)
}
}
