% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_fit.R
\name{plot_subgroup_ates}
\alias{plot_subgroup_ates}
\title{Plot ATE differences by one or more subgroups}
\usage{
plot_subgroup_ates(fit, subgroups, level = 0.9)
}
\arguments{
\item{fit}{A trained causal forest object from
\code{\link[grf]{causal_forest}}}

\item{subgroups}{A logical vector indicating whether an observation belongs
to the subgrouping, or a list/data frame of such vectors.}

\item{level}{The width of the reported confidence intervals}
}
\value{
A ggplot2 plot object
}
\description{
Visualize the heterogeneity heuristic described in
\code{\link{estimate_subgroup_ate}}. If more than one subgroup is given, they
are all displayed in the plot.
}
\details{
If \code{subgroups} is a \code{data.frame}, the columns are assumed to be the
subgroups. They should all be logical, otherwise unexpected behavior may
occur.
}
\examples{
\dontrun{
 require(grf)

 n <- 2000; p <- 10

 X <- matrix(rnorm(n * p), n, p)
 W <- rbinom(n, 1, 0.4 + 0.2 * (X[, 1] > 0))
 Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)

 cf <- causal_forest(X, Y, W)
 results <- tidy_cf(cf)

 grps <- data.frame(
   high_cate = results$cate > median(results$cate),
   high_X3 = X[, 3] > median(X[, 3])
 )
 plot_subgroup_ates(cf, grps)
}
}
