% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Create a causal forest variable importance object}
\usage{
varimp(fit)
}
\arguments{
\item{fit}{A trained \code{\link[grf]{causal_forest}} object}
}
\value{
An object of class \code{varimp}, essentially just a
\link[tibble:tibble-package]{tibble} with a row for each covariate.
}
\description{
Create a causal forest variable importance object
}
\examples{
\dontrun{
 require(grf)

 Xdat <- subset(cfex, select = -c(W, Y))
 X <- make_contrasts(Xdat, 'fct')
 cf <- causal_forest(X, cfex$Y, cfex$W)

 cf_eval(cf, Xdat)$varimp
}
}
