% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFscore.R
\name{CFscore}
\alias{CFscore}
\title{Assess counterfactual performance of a model capable of predictions under
interventions}
\usage{
CFscore(data, model, Y, propensity_formula, treatments)
}
\arguments{
\item{data}{A data.frame on which the model is to be validated.}

\item{model}{A glm (lm?) which can make predictions under interventions.}

\item{Y}{A character string indicating the name of the observed outcome
column in data, or a numeric vector of the observed outcomes.}

\item{propensity_formula}{A formula used to estimate the inverse-probability
weights for the validation data. Treatment variable should be on the left
hand side, all confounders on the right hand side. It is possible that
there is a different set of confounders in the validation dataset compared
to the model-development dataset.}

\item{treatments}{A list of all treatments for which the counterfactual
perormance measures should be evaluated.}
}
\value{
A list of performance measures (Brier, Observed/Expected, AUC) of
  the model on observed data (naive) and counterfactual data, where each
  subject is assigned to both treatment options.
}
\description{
Assess counterfactual performance of a model capable of predictions under
interventions
}
\examples{
causal_model <- glm(
  Y ~ A + P,
  family = "binomial",
  data = df_dev,
  weights = ip_weights(df_dev, A ~ L)
)
CFscore(df_val, causal_model, "Y", A ~ L, treatments = list(0,1))
}
