% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFscore.R
\name{CFscore}
\alias{CFscore}
\title{Assess counterfactual performance of a model capable of predictions under
interventions}
\usage{
CFscore(
  data,
  model,
  Y_column_name,
  propensity_formula,
  treatments,
  quiet_mode = FALSE
)
}
\arguments{
\item{data}{A data.frame on which the model is to be validated.}

\item{model}{A glm (lm?) which can make predictions under interventions.}

\item{Y_column_name}{The observed outcome column of the data.}

\item{propensity_formula}{A formula used to estimate the inverse-probability
weights for the validation data. Treatment variable should be on the left
hand side, all confounders on the right hand side. It is possible that
there is a different set of confounders in the validation dataset compared
to the model-development dataset.}

\item{quiet_mode}{Set to TRUE to avoid printing all assumptions.}
}
\value{
A list of performance measures (Brier, Observed/Expected, AUC) of
  the model on observed data (naive) and counterfactual data, where each
  subject is assigned to both treatment options.
}
\description{
Assess counterfactual performance of a model capable of predictions under
interventions
}
\examples{
causal_model <- glm(
  Y ~ A + P,
  family = "binomial",
  data = df_dev,
  weights = ip_weights(df_dev, A ~ L)
)
df_val <- build_data(2000)
CFscore(df_val, causal_model, "Y", A ~ L)
}
